% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{abx_sensitive}
\alias{abx_sensitive}
\title{Checks whether a bacteria has a specific antibiotic sensitivity}
\usage{
abx_sensitive(strainData, abx = "vancomycin")
}
\arguments{
\item{strainData}{A list from the function \code{\link{getStrainData}}}

\item{abx}{A specific antibiotic e.g. "vancomycin"}
}
\value{
Logical TRUE or FALSE
}
\description{
Checks whether a bacteria has a specific antibiotic sensitivity
}
