% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{abxResistant}
\alias{abxResistant}
\title{Checks whether a bacteria has a specific antibiotic resistance}
\usage{
abxResistant(strainData, abx = "vancomycin")
}
\arguments{
\item{strainData}{A list from the function \code{\link{getStrainData}}}

\item{abx}{A specific antibiotic e.g. "vancomycin"}
}
\value{
TRUE, FALSE or NULL (antibiotic info is not found)
}
\description{
Checks whether a bacteria has a specific antibiotic resistance
}
\examples{
\dontrun{
strain_list <- getStrains(page = 1,
genus = 'Pseudomonas',
species = 'tarimensis',
userpassword = paste0(user,':',passwd))
phenotypes_list <- getStrainData(strain_list,
selection=1,
userpassword = paste0(user,':',passwd))
abxResistant(phenotypes_list, abx = "gentamicin")
}
}
