% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{abxSensitive}
\alias{abxSensitive}
\title{Checks whether a bacteria has a specific antibiotic sensitivity}
\usage{
abxSensitive(strainData, abx = "vancomycin")
}
\arguments{
\item{strainData}{A list from the function \code{\link{getStrainData}}}

\item{abx}{A specific antibiotic e.g. "vancomycin"}
}
\value{
TRUE, FALSE or NULL (antibiotic info is not found)
}
\description{
Checks whether a bacteria has a specific antibiotic sensitivity
}
