% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_abx}
\alias{bacdive_abx}
\title{Antibiotics data from bacdive.org that is _not_ in the antibiogram format. Produced by \code{\link{getAbx}}.}
\format{
A data frame with the following columns:
\describe{
  \item{`@ref`}{The literature reference in the bacdive database.}
  \item{ChEBI}{\href{https://www.ebi.ac.uk/chebi/}{ChEBI (Chemical entities of biological interest) identifier.}}
  \item{metabolite}{Common name of the antibiotic.}
  \item{`is antibiotic`}{Should all be "yes" values.}
  \item{`is sensitive`}{Whether or not strain is sensitive to this antibiotic.
  Possible values are "yes", "no", and NA.}
  \item{`is resistant`}{Whether or not strain is resistant to this antibiotic.
  Possible values are "yes", "no", and NA.}
  \item{ID}{bacdive ID of strain.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{type_strain}{Whether or not the strain is the type strain for the species.}
  \item{`doi/url`}{DOI (Digital Object Identifier) for the reference.}
  \item{`resistance conc.`}{The concentration of antibiotic at which resistance was measured.}
  \item{`sensitivity conc.`}{The concentration of antibiotic at which sensitivity was measured.}
  \item{`group ID`}{Group ID of combined antibiotics
  i.e. bacteria was tested against multiple antibiotics at once.}
  \item{`is intermediate`}{Whether or not strain is intermediate sensitive to this antibiotic.}
  \item{`intermediate conc.`}{The concentration of antibiotic at which intermediate sensistivity was measured.}
}
}
\usage{
bacdive_abx
}
\description{
Antibiotics data from bacdive.org that is _not_ in the antibiogram format. Produced by \code{\link{getAbx}}.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
