% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_antibiogram}
\alias{bacdive_antibiogram}
\title{Antibiotics data from bacdive.org that is in the \href{https://bacdive.dsmz.de/help/what-are-antibiotic-susceptibi.htm}antibiogram format. Produced by \code{\link{getAntibiogram}}.}
\format{
A data frame with the following columns:
\describe{
  \item{`@ref`}{The literature reference in the bacdive database.}
  \item{medium}{Growth medium used for sensitivity / resistance test.}
  \item{`incubation temperature`}{Incubation temp of the test in degrees Celsius.}
  \item{`oxygen condition`}{Oxygen exposure during the test.}
  \item{metabolite}{Common name of the antibiotic.}
  \item{diameter}{Inhibition zone diameter in mm.}
  \item{ID}{bacdive ID of strain.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{type_strain}{Whether or not the strain is the type strain for the species.}
  \item{`doi/url`}{DOI (Digital Object Identifier) for the reference.}
  \item{`incubation time`}{Incubation time of test in days.}
}
}
\usage{
bacdive_antibiogram
}
\description{
Antibiotics data from bacdive.org that is in the \href{https://bacdive.dsmz.de/help/what-are-antibiotic-susceptibi.htm}antibiogram format. Produced by \code{\link{getAntibiogram}}.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
