% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_enzymes}
\alias{bacdive_enzymes}
\title{Enzyme activity data from bacdive.org. Produced by \code{\link{getEnzymes}}.}
\format{
A data frame with the following columns:
\describe{
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{value}{The enzyme name.}
  \item{ec}{\href{https://en.wikipedia.org/wiki/Enzyme_Commission_number}{The Enzyme Commission number.}}
  \item{activity}{Whether the enzyme is active or not with '+' being active, '-' being non-active, and '+/-' being variable. By default, NA values are removed and the most common value is shown for a given species
  if it has multiple strains.}
  }
}
\usage{
bacdive_enzymes
}
\description{
Enzyme activity data from bacdive.org. Produced by \code{\link{getEnzymes}}.
}
\details{
See `data-raw/download_from_bacdive.Rmd` for code to create this.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
