% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_morphology}
\alias{bacdive_morphology}
\title{Cell morphology data from bacdive.org. Produced by \code{\link{getMorphology}}.}
\format{
A data frame with the following columns:
\describe{
  \item{`@ref`}{The literature reference in the bacdive database.}
  \item{`gram stain`}{How the taxon appears when Gram-stained.
  Multiple values including "positive", "negative" or "variable". NA means the status is unknown.}
  \item{`cell length`}{Cell length in micrometers.}
  \item{`cell width`}{Cell width in micrometers.}
  \item{`cell shape`}{Cell shape.}
  \item{motility}{Whether or not strain can move.}
  \item{`flagellum arrangement`}{Where the flagella are arranged on the cell.}
  \item{ID}{bacdive ID of strain.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{type_strain}{Whether or not the strain is the type strain for the species.}
  \item{`doi/url`}{DOI (Digital Object Identifier) for the reference.}
}
}
\usage{
bacdive_morphology
}
\description{
Cell morphology data from bacdive.org. Produced by \code{\link{getMorphology}}.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
