% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_oxygen}
\alias{bacdive_oxygen}
\title{Oxygen tolerance data from bacdive.org. Produced by \code{\link{getOxygen}}.}
\format{
A data frame with the following columns:
\describe{
  \item{`@ref`}{The literature reference in the bacdive database.}
  \item{`oxygen tolerance`}{The oxygen tolerance of the strain.
  Multiple values including "aerobe", "facultative anaerobe", and
    "obligate anaerobe", etc. NA means the status is unknown or variable.}
  \item{ID}{bacdive ID of strain.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{type_strain}{Whether or not the strain is the type strain for the species.}
  \item{`doi/url`}{DOI (Digital Object Identifier) for the reference.}
}
}
\usage{
bacdive_oxygen
}
\description{
Oxygen tolerance data from bacdive.org. Produced by \code{\link{getOxygen}}.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
