% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_phenotypes}
\alias{bacdive_phenotypes}
\title{Phenotype data of bacterial type strains from bacdive.org}
\format{
A data frame with the following columns:
\describe{
  \item{ID}{The BacDive ID.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{aerobic_status}{
    The aerobic status. Multiple values including "aerobe", "facultative anaerobe", and
    "obligate anaerobe".}
  \item{gram_stain}{
    How the taxon appears when Gram-stained. One of "positive" or
    "negative".}
}
}
\usage{
bacdive_phenotypes
}
\description{
Phenotype data of bacterial type strains from bacdive.org
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
