% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{getAbxOld}
\alias{getAbxOld}
\title{Get full antibiotic susceptibility / resistance information}
\usage{
getAbxOld(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of antibiotic susceptibility / resistance information about taxa in the list.


}
\description{
Get full antibiotic susceptibility / resistance information
}
\examples{
\dontrun{
bacdive_abx <- getAbx(list_holder)
}
}
