% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getAbxSingle}
\alias{getAbxSingle}
\title{Get the regular antibiotic susceptibility / resistance information from a single bacdive entry}
\usage{
getAbxSingle(bacdive_entry)
}
\arguments{
\item{bacdive_entry}{A single entry representing a strain within a bacdive list.}
}
\value{
A dataframe of antibiotic susceptibility / resistance information for a single bacdive entry.
}
\description{
Get the regular antibiotic susceptibility / resistance information from a single bacdive entry
}
\details{
Some antibiotic information is encoded as an antibiogram, see \code{\link{getAntibiogramSingle}}
}
\examples{
\dontrun{
abx_df <- bind_rows(lapply(list_holder, getAbxSingle))
# OR
single_strain <- getAbxSingle(list_holder[[1]])
}
}
