% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getAntibiogram}
\alias{getAntibiogram}
\title{Get the antibiogram susceptibility / resistance information from a list of bacdive entries}
\usage{
getAntibiogram(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of antibiogram susceptibility / resistance information about taxa in the list.
}
\description{
Get the antibiogram susceptibility / resistance information from a list of bacdive entries
}
\examples{
\dontrun{
antibiogram_df <- getAntibiogram(list_holder)
}
}
