% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getAntibiogramSingle}
\alias{getAntibiogramSingle}
\title{Get the antibiogram susceptibility / resistance information for a single entry}
\usage{
getAntibiogramSingle(bacdive_entry)
}
\arguments{
\item{bacdive_entry}{A single entry representing a strain within a bacdive list.}
}
\value{
A dataframe of antibiogram susceptibility / resistance information for a single bacdive entry.
}
\description{
Get the antibiogram susceptibility / resistance information for a single entry
}
\details{
This is similar information from what is gotten from \code{\link{getAbxSingle}} but uses a standardized format called the antibiogram
}
\examples{
\dontrun{
antibiogram_df <- bind_rows(lapply(list_holder, getAntibiogramSingle))
# OR
single_strain <- getAntibiogramSingle(list_holder[[1]])
}
}
