% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_strain_access.R
\name{getBacDiveAccess}
\alias{getBacDiveAccess}
\title{Opens a portal to the world of BacDive}
\usage{
getBacDiveAccess(
  credentials = Sys.getenv(c("DSMZ_API_USER", "DSMZ_API_PASSWORD"))
)
}
\arguments{
\item{credentials}{Your login credentials for bacdive.org. Defaults to DSMZ_API_USER and DSMZ_API_PASSWORD in your Renviron file. Supplied as a character vector c("username", "password").}
}
\value{
A 'dsmz_keycloak' object that allows access to the BacDive API. See also \code{\link[BacDive]{open_bacdive}}.
}
\description{
Opens a portal to the world of BacDive
}
\examples{
\dontrun{
bacdive <- getBacDiveAccess()
}
}
