% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getEnzymes}
\alias{getEnzymes}
\title{Extracts enzyme data from a bacdive list}
\usage{
getEnzymes(list_holder = list_holder, most_common = T, remove_unknown = T)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}

\item{most_common}{The function will take the most common value for any given species enzyme activity combination. See details.}

\item{remove_unknown}{Remove enzyme activity entries where the value is NA.}
}
\value{
A dataframe of enzyme information about taxa in the list.
}
\description{
Extracts enzyme data from a bacdive list
}
\details{
Staphylococcus aureus has 31 strains with positive catalase activity and 2 strains with negative catalase activity. With `most_common = T`, the function will return a dataframe denoting that S. aureus is most commonly positive for catalase activity.
}
\examples{
\dontrun{
bacdive_enzymes <- getEnzymes(list_holder)
}
}
