% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getEnzymesSingle}
\alias{getEnzymesSingle}
\title{Extracts enzyme data from a bacdive entry}
\usage{
getEnzymesSingle(bacdive_entry)
}
\arguments{
\item{bacdive_entry}{A single entry representing a strain within a bacdive list.}
}
\value{
A dataframe of enzyme information for a single bacdive entry.
}
\description{
Extracts enzyme data from a bacdive entry
}
\examples{
\dontrun{
enzymes_df <- bind_rows(lapply(list_holder, getEnzymesSingle))
# OR
single_strain <- getEnzymesSingle(list_holder[[1]])
}
}
