% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{getMorphologyOld}
\alias{getMorphologyOld}
\title{Obsolete -- do no use -- for testing porpoises only}
\usage{
getMorphologyOld(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of cell morphology information about taxa in the list.
}
\description{
Obsolete -- do no use -- for testing porpoises only
}
\examples{
\dontrun{
morphology_df <- getMorphologyOld(list_holder)
}
}
