% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getMorphologySingle}
\alias{getMorphologySingle}
\title{Extracts cell morphology data from a single entry in a bacdive structured list}
\usage{
getMorphologySingle(bacdive_entry)
}
\arguments{
\item{bacdive_entry}{A single entry representing a strain within a bacdive list.}
}
\value{
A dataframe of cell morphology information about the strain.
}
\description{
Extracts cell morphology data from a single entry in a bacdive structured list
}
\note{
Used in \code{\link{getMorphology}}.
}
\examples{
\dontrun{
morphology_df <- bind_rows(lapply(list_holder, getMorphologySingle))
# OR
single_strain <- getMorphologySingle(list_holder[[1]])
}
}
