% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getOxygen}
\alias{getOxygen}
\title{Extracts oxygen tolerance data from a bacdive list}
\usage{
getOxygen(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of oxygen tolerance information about taxa in the list.
}
\description{
Extracts oxygen tolerance data from a bacdive list
}
\examples{
\dontrun{
oxygen_df <- getOxygen(list_holder)
}
}
