% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getOxygenSingle}
\alias{getOxygenSingle}
\title{Extracts oxygen tolerance data from a single entry in a bacdive structured list}
\usage{
getOxygenSingle(bacdive_entry)
}
\arguments{
\item{bacdive_entry}{A single entry representing a strain within a bacdive list.}
}
\value{
A dataframe of oxygen tolerance information about the strain.
}
\description{
Extracts oxygen tolerance data from a single entry in a bacdive structured list
}
\note{
Used in \code{\link{getOxygen}}.
}
\examples{
\dontrun{
morphology_df <- bind_rows(lapply(list_holder, getOxygenSingle))
# OR
single_strain <- getOxygenSingle(list_holder[[1]])
}
}
