% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Combined oxygen tolerance and gram stain data for use with other packages}
\usage{
getPhenotypes(morphology_df, oxygen_df)
}
\arguments{
\item{morphology_df}{A dataframe from \code{\link{getMorphology}}.}

\item{oxygen_df}{A dataframe from \code{\link{getOxygen}}.}
}
\value{
A dataframe of oxygen tolerance and gram stain information about taxa in the list.
}
\description{
Combined oxygen tolerance and gram stain data for use with other packages
}
\examples{
\dontrun{
morphology_df <- getMorphology(list_holder)
oxygen_df <- getOxygen(list_holder)
phenotype_df <- getPhenotypes(morphology_df, oxygen_df)
}
}
