% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getSimplifiedAbx}
\alias{getSimplifiedAbx}
\title{Simplify the full antibiotic susceptibility / resistance information}
\usage{
getSimplifiedAbx(
  data = bacdive_abx,
  extra_info = F,
  most_common = T,
  remove_unknown = T
)
}
\arguments{
\item{data}{A dataframe of antibiotic susceptibility / resistance information from \code{\link{getAbx}}.}

\item{extra_info}{Whether to return some extra info for testing purposes. This will include the columns `is resistant`, `is_sensitive`, and `diameter`.}

\item{most_common}{The function will take the most common value for any given antibiotic / species combination. See details.}

\item{remove_unknown}{Remove antibiotic / species combinations where the resistance / sensitivity information is unknown or variable.}
}
\value{
A simplified dataframe of antibiotic susceptibility / resistance information.
}
\description{
Simplify the full antibiotic susceptibility / resistance information
}
\details{
Hypothetically, a species might have 50 strains that are resistant to Penicillin and 25 strains that are sensitive. With `most_common = T` the function will collapse the results to the most common value. Thus, the returned dataframe will have this species as being resistant to Penicillin.
}
\examples{
\dontrun{
bacdive_susceptibility <- getSimplifiedAbx(bacdive_abx)
}
}
