% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_strain_access.R
\name{getStrainLocal}
\alias{getStrainLocal}
\title{Gets a single strain of a species}
\usage{
getStrainLocal(
  list = NULL,
  rda = "data-raw/strain_large_list.rda",
  query,
  typestrain_only = F
)
}
\arguments{
\item{list}{A list object containing strain information already present in the R environment. Required if an rda file is not given.}

\item{rda}{An RData file representing a full download of strain from BacDive.org's API. Can be created using download_from_bacdive.Rmd. Required if list is not given.}

\item{query}{Either a bacdive ID (integer) or species name (character).}

\item{typestrain_only}{Whether to filter down to typestrains only.}
}
\value{
A list containing information on the strain of a species.
}
\description{
Gets a single strain of a species
}
\details{
If you give typestrain_only = T and your BacDive-ID query is not a typestrain you will get zero results.
}
\examples{
\dontrun{
abyss <- getStrainLocal(query = "Abyssibacter profundi")
}
}
