% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getStrains}
\alias{getStrains}
\title{Searching for strains with taxon specification}
\usage{
getStrains(
  credentials = Sys.getenv(c("DSMZ_API_USER", "DSMZ_API_PASSWORD")),
  query,
  sleep = 0.1,
  handler = NULL,
  getList = BacDive::retrieve
)
}
\arguments{
\item{credentials}{Your login credentials for bacdive.org. Defaults to DSMZ_API_USER and DSMZ_API_PASSWORD in your Renviron file. Supplied as a character vector c("username", "password").}

\item{sleep}{A waiting period in seconds between successive API requests, if any.}
}
\value{
A list of strains with phenotype information from bacdive.org
}
\description{
Searching for strains with taxon specification
}
\examples{
\dontrun{
strain_list <- getStrains(query = "Bacteroides xylanisolvens",
sleep = 0.1)
}
}
