% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getTypeStrainLocal}
\alias{getTypeStrainLocal}
\title{Gets a single type strain of a species}
\usage{
getTypeStrainLocal(
  list = NULL,
  rda = "data-raw/type_strain_large_list.rda",
  query
)
}
\arguments{
\item{list}{A list object containing strain information already present in the R environment. Required if an rda file is not given.}

\item{rda}{An RData file representing a full download of type strain from BacDive.org's API. Can be created using download_from_bacdive.Rmd. Required if list is not given.}

\item{query}{A species.}
}
\value{
A list containing information on the type strain of a species.
}
\description{
Gets a single type strain of a species
}
\examples{
\dontrun{
abyss <- getTypeStrainLocal(query = "Abyssibacter profundi")
}
}
