% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{oxygenTolerance}
\alias{oxygenTolerance}
\title{Gets the oxygen tolerance of the strain}
\usage{
oxygenTolerance(strainData, reference = 1)
}
\arguments{
\item{strainData}{A list from the function \code{\link{getStrainData}}}

\item{reference}{Sometimes strains have multiple references for the phenotype information,
this is where you can select the certain reference (or iterate through them)}
}
\value{
A character string e.g. "aerobe", "anaerobe", etc.
}
\description{
Gets the oxygen tolerance of the strain
}
