% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Shen2021}
\alias{Shen2021}
\title{Taxonomic count data from shotgun metagenomics}
\format{
A data frame with the following columns:
\describe{
  \item{SampleID}{The unique SampleID.}
  \item{SampleType}{The type of sample: either "Rectal swab" or "Feces".}
  \item{ETOH_etiology}{Whether the liver disease was alcohol related or not.}
  \item{Antibiotics}{The type of antibiotic the patient was taking.}
  \item{MELD}{MELD Score (Model For End-Stage Liver Disease). Higher numbers indicate worse prognosis.}
  \item{read_counts}{Total read counts in the sample.}
  \item{taxon}{Taxon at the species level.}
  \item{count}{Count of the Taxon.}
}
}
\usage{
Shen2021
}
\description{
Taxonomic count data from shotgun metagenomics
}
\note{
Data obtained from the following study: Shen, T.-C. D. et al. (2021) ‘The Mucosally-Adherent Rectal Microbiota Contains Features Unique to Alcohol-Related Cirrhosis’, Gut microbes, 13(1), p. 1987781. \url{https://www.ncbi.nlm.nih.gov/pubmed/34747331}.
}
\keyword{datasets}
