% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_enzymes}
\alias{bacdive_enzymes}
\title{Enzyme activity data from bacdive.org}
\format{
A data frame with the following columns:
\describe{
  \item{ID}{The BacDive ID.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{activity}{Whether the enzyme is active or not with '+' being active, '-' being non-active, and '+/-' being variable.}
  \item{value}{The enzyme name.}
  \item{ec}{\href{https://en.wikipedia.org/wiki/Enzyme_Commission_number}{The Enzyme Commission number.}}
  \item{doi}{Digital object identifier for the literature that supports the evidence of activity.}
}
}
\usage{
bacdive_enzymes
}
\description{
Enzyme activity data from bacdive.org
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
