% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_phenotypes}
\alias{bacdive_phenotypes}
\title{Phenotype data of bacterial type strains from bacdive.org.  Produced by \code{\link{getPhenotypes}}.}
\format{
A data frame with the following columns:
\describe{
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{aerobic_status}{The aerobic status.
  Multiple values including "aerobe", "facultative anaerobe", and
    "obligate anaerobe", etc. NA means the status is unknown or variable.}
  \item{gram_stain}{How the taxon appears when Gram-stained.
  Multiple values including "positive", "negative" or "variable". NA means the status is unknown.}
}
}
\usage{
bacdive_phenotypes
}
\description{
Phenotype data of bacterial type strains from bacdive.org.  Produced by \code{\link{getPhenotypes}}.
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
