% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacdive_susceptibility}
\alias{bacdive_susceptibility}
\title{Antibiotic susceptibility data from bacdive.org}
\format{
A data frame with the following columns:
\describe{
  \item{ID}{The BacDive ID.}
  \item{taxon}{The name of the taxon.}
  \item{rank}{The rank of the taxon.}
  \item{antibiotic}{The antibiotic or antibiotic class.}
  \item{value}{
    The susceptibility of the taxon to the antibiotic, one of "susceptible"
    or "resistant".}
}
}
\usage{
bacdive_susceptibility
}
\description{
Antibiotic susceptibility data from bacdive.org
}
\note{
Data frame also has the attribute "date_downloaded" for when the data was downloaded from bacdive.org.
}
\keyword{datasets}
