% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_extract.R
\name{getAbx}
\alias{getAbx}
\title{Get the regular antibiotic susceptibility / resistance information from a list of bacdive entries}
\usage{
getAbx(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of antibiotic susceptibility / resistance information about taxa in the list.
}
\description{
Get the regular antibiotic susceptibility / resistance information from a list of bacdive entries
}
\examples{
\dontrun{
abx_df <- getAbx(list_holder)
}
}
