% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacPhene.R
\name{getMorphology}
\alias{getMorphology}
\title{Extracts cell morphology data from a bacdive list}
\usage{
getMorphology(list_holder = list_holder)
}
\arguments{
\item{list_holder}{A list object containing strain information already present in the R environment.}
}
\value{
A dataframe of cell morphology information about taxa in the list.
}
\description{
Extracts cell morphology data from a bacdive list
}
\details{
Essentially a wrapper for applying \code{\link{getMorphologySingle}} to a list and turning the result into a dataframe. Also adds the "date_downloaded" attribute from the list to the dataframe.
}
\examples{
\dontrun{
morphology_df <- getMorphology(list_holder)
}
}
