% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacdive_strain_access.R
\name{getStrains}
\alias{getStrains}
\title{Downloads BacDive strains from a list of supplied ID's}
\usage{
getStrains(
  strain_list = "data-raw/full_list_of_bacteria_from_bacdive_20211027.csv",
  typestrain_only = T,
  bacdive_keycloak = getBacDiveAccess(),
  rda = "data-raw/strain_large_list.rda"
)
}
\arguments{
\item{strain_list}{A csv file containing at least two columns: "ID" - The BacDive ID and "is_type_strain_header". Can be gotten using \href{https://bacdive.dsmz.de/advsearch}{BacDive's advanced search}.}

\item{typestrain_only}{Whether to filter down to typestrains only.}

\item{bacdive_keycloak}{A 'dsmz_keycloak' object that allows access to the BacDive API. Created using \code{\link{getBacDiveAccess}} or \code{\link[BacDive]{open_bacdive}}.}

\item{save_rda}{If this is being run for the first time, the list of strains will be saved to the specificed RData file. Otherwise, the function will load strains from the specified data file.}
}
\value{
A list object of strain information. If running for the first time, saves an RData file of the list.
}
\description{
Downloads BacDive strains from a list of supplied ID's
}
\examples{
\dontrun{
large_list <- getStrains(strain_list = "my_csv_file.csv", typestrain_only = F)
}
}
